Move Java tests outside spark package
