+++ src/test/java/uk/co/gresearch/test/diff/JavaValue.java
