+++ src/test/java/uk/co/gresearch/test/diff/JavaValueAs.java
